pyproject.toml
[tool.poetry]
name = "django-react-chat"
version = "0.1.0"
description = ""
authors = ["vaduska <vadimsalapugin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
python-dotenv = "^0.21.0"
django-rest-framework = "^0.1.0"
PyJWT = "^2.6.0"
Pillow = "^9.4.0"
requests = "^2.28.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
